Improve exception and refactoring

+++ app/Services/TypeTodoService.php
@@ -1,8 +1,14 @@
 <?php
 
 namespace App\Services;
+use App\Services\TodoService;
 
 use App\Repositories\TypeTodoRepository;
+
+use App\Exceptions\TypeNotExistsException;
+use App\Exceptions\TypeAlreadyExistsException;
+use App\Exceptions\TypeAlreadyInUseException;
+
 
 class TypeTodoService
 {
@@ -17,10 +23,12 @@ class TypeTodoService
     {
         return $this->respository->getAll();
     }
-    
 
     public function create(array $data)
     {
+        if($this->exists($data['description'])){
+            throw new TypeAlreadyExistsException('Tipo de tarefa já existe.');
+        }
         return $this->respository->create($data);
     }
 
@@ -31,11 +39,22 @@ class TypeTodoService
 
     public function update(array $data, int $id)
     {
+        $typesByDescriptions = $this->getByDescription($array['description']);
+        if(isset($typesByDescriptions[0]) && $typesByDescriptions[0]->id != $id){
+            throw new TypeAlreadyExistsException('Tipo de tarefa já existe.');
+        }
         return $this->respository->update($data, $id);
     }
 
     public function delete(int $id)
     {
+        if($this->get($id) == null){
+            throw new TypeNotExistsException('Tipo não existe.');
+        }
+        if((new TodoService)->existsTodoByType($id)){
+            throw new TypeAlreadyInUseException('Não é possível deletar um Tipo em uso.');
+        }
+
         return $this->respository->delete($id);
     }
 
@@ -43,7 +62,7 @@ class TypeTodoService
     {
         return $this->respository->getByDescription($description);
     }
-    
+
     public function exists(string $description)
     {
         return count($this->respository->getByDescription($description)) > 0;
